exclude partitionedBy column at Projection stage
instead of at InsertSourceGen, same as how COPY handles it
